+++ apps/RecordThing/Shared/Navigation/AppSplitView.swift
@@ -53,6 +53,98 @@ struct AppSplitView<DetailContent: View>: View {
     self.detailContent = detailContent()
   }
 
+  #if os(macOS)
+    var macosList: some View {
+      List(selection: $lifecycleView) {
+        NavigationLink(value: LifecycleView.record) {
+          Label("Record", systemImage: "camera")
+        }
+        .accessibilityLabel("Record")
+
+        NavigationLink(value: LifecycleView.assets) {
+          Label("Assets", systemImage: "square.3.layers.3d")
+        }
+        .accessibilityLabel("Assets")
+
+        NavigationLink(value: LifecycleView.actions) {
+          Label("Actions", systemImage: "signature")
+        }
+        .accessibilityLabel("Actions")
+
+        NavigationLink(value: LifecycleView.settings) {
+          Label("Settings", systemImage: "gear")
+        }
+        .accessibilityLabel("Settings")
+
+        NavigationLink {
+          ThingsMenu()
+        } label: {
+          Label("Things", systemImage: "square.3.layers.3d")
+        }
+
+        NavigationLink {
+          EvidenceTypeMenu()
+        } label: {
+          Label("Evidence Type", systemImage: "square.3.layers.3d")
+        }
+
+        Section("Actions") {
+          Button(action: {
+            evidenceViewModel.reviewing.toggle()
+          }) {
+            Label("Review Evidence", systemImage: "photo.stack")
+          }
+        }
+      }
+      .listStyle(.sidebar)
+    }
+  #endif
+
+  var iosList: some View {
+    List {
+      Button(action: { lifecycleView = .record }) {
+        Label("Record", systemImage: "camera")
+      }
+      .accessibilityLabel("Record")
+
+      Button(action: { lifecycleView = .assets }) {
+        Label("Assets", systemImage: "square.3.layers.3d")
+      }
+      .accessibilityLabel("Assets")
+
+      Button(action: { lifecycleView = .actions }) {
+        Label("Actions", systemImage: "signature")
+      }
+      .accessibilityLabel("Actions")
+
+      Button(action: { lifecycleView = .settings }) {
+        Label("Settings", systemImage: "gear")
+      }
+      .accessibilityLabel("Settings")
+
+      NavigationLink {
+        ThingsMenu()
+      } label: {
+        Label("Things", systemImage: "square.3.layers.3d")
+      }
+
+      NavigationLink {
+        EvidenceTypeMenu()
+      } label: {
+        Label("Evidence Type", systemImage: "square.3.layers.3d")
+      }
+
+      Section("Actions") {
+        Button(action: {
+          evidenceViewModel.reviewing.toggle()
+        }) {
+          Label("Review Evidence", systemImage: "photo.stack")
+        }
+      }
+    }
+    .listStyle(.sidebar)
+  }
+
   var body: some View {
     NavigationSplitView(
       columnVisibility: $columnVisibility, preferredCompactColumn: $preferredColumn
@@ -77,131 +169,15 @@ struct AppSplitView<DetailContent: View>: View {
 
         // Blur effect
         //                VisualEffectView()
-        List {
-
-          // This updates content
-          NavigationLink {
-            if let assetsViewModel = assetsViewModel {
-              ThingsGridView(
-                viewModel: assetsViewModel, evidenceViewModel: evidenceViewModel, columns: 2
-              ) { thing in
-                print("Selected thing: \(thing.title ?? "Untitled")")
-              }
-            } else {
-              Text("Missing Assets")
-            }
-          } label: {
-            Label("Assets", systemImage: "square.3.layers.3d")
-          }
-
-          NavigationLink {
-            List {
-              Section {
-                NavigationLink("Settings") {
-                  ImprovedSettingsView(
-                    captureService: captureService,
-                    designSystem: designSystem
-                  )
-                }
-
-                Button("Update Account") {
-                  // Account update action
-                }
-
-              } header: {
-                Text("Actions")
-              }
-
-              Section {
-                Text("Account Profile")
-              } header: {
-                Text("Account & Teams")
-              }
-            }
-            .navigationTitle("Actions")
-          } label: {
-            Label("Actions", systemImage: "signature")
-          }
-
-          NavigationLink {
-            ImprovedSettingsView(
-              captureService: captureService,
-              designSystem: designSystem
-            )
-          } label: {
-            Label("Settings", systemImage: "gear")
-          }
-
-          NavigationLink {
-            ThingsMenu()
-          } label: {
-            Label("Things", systemImage: "square.3.layers.3d")
-          }
-
-          NavigationLink {
-            EvidenceTypeMenu()
-          } label: {
-            Label("Evidence Type", systemImage: "square.3.layers.3d")
-          }
-
-          //                    Section("Recent Evidence") {
-          //                        ForEach(evidenceViewModel.pieces) { evidence in
-          //                            EvidenceListItem(evidence: evidence)
-          //                        }
-          //                    }
-
-          Section("Actions") {
-            Button(action: {
-              evidenceViewModel.reviewing.toggle()
-            }) {
-              Label("Review Evidence", systemImage: "photo.stack")
-            }
-          }
-
-        }
-        .listStyle(.sidebar)
+        #if os(macOS)
+          macosList
+        #else
+          iosList
+        #endif
       }
       .navigationTitle("Record Thing")
-      .toolbar {
-        ToolbarItem {
-          NavigationLink {
-            detailContent.onAppear {
-              lifecycleView = .record
-            }
-          } label: {
-            Label("Record", systemImage: "camera")
-          }
-        }
-
-      }
-      //        } content: {
-      //            // blank content
-      //            Button("Detail Only") {
-      //                columnVisibility = .detailOnly
-      //            }
-      //            Button("Double Column") {
-      //                columnVisibility = .doubleColumn
-      //            }
-      //            NavigationLink {
-      //                Text("sdfsdfsafd (detail)")
-      //            } label: {
-      //                Label("123 (D)", systemImage: "gear")
-      //            }
-      //
     } detail: {
-      NavigationStack {
-        VStack {
-          //                    NavigationLink {
-          //                        Text("sdfsdfsafd (detail)")
-          //                    } label: {
-          //                        Label("123 (D)", systemImage: "gear")
-          //                    }
-          detailContent
-        }
-        .navigationDestination(for: NavigationDestination.self) { dest in
-          AppSplitDetailView(destination: dest)
-        }
-      }
+      detailContent
     }
     .navigationSplitViewStyle(.balanced)
     #if os(macOS)
@@ -225,79 +201,6 @@ struct AppSplitView<DetailContent: View>: View {
          */
       }
     #endif
-  }
-}
-
-struct AppSplitDetailView: View {
-  let destination: NavigationDestination
-
-  var body: some View {
-    Group {
-      switch destination {
-      case .record(let evidenceId):
-        EmptyView()
-      case .assets:
-        ZStack {
-          Text("Assets")
-        }
-      case .assetsGroup(let groupId):
-        EmptyView()
-      case .assetsDetail(let assetId):
-        EmptyView()
-      case .assetsThingDetail(let thingId):
-        EmptyView()
-      case .assetsEvidenceList(let thingId):
-        EmptyView()
-      case .assetsThingDetail(let thingId):
-        EmptyView()
-      case .assetsEvidenceDetail(let thingId, let evidenceId):
-        EmptyView()
-
-      // Actions Tab Paths
-      case .actions:
-        ZStack {
-          Text("Actions")
-        }
-      case .actionsAccount:
-        EmptyView()
-      case .actionsSettings:
-        EmptyView()
-      case .actionsHelp:
-        EmptyView()
-      }
-
-      /*
-      switch destination {
-      case .thingDetail(let id):
-          if let thing = sampleThings.first(where: { $0.id == id }) {
-              ThingDetailView(thing: thing)
-          }
-      case .categoryDetail(let id):
-          CategoryDetailView(categoryId: id)
-      case .productTypeDetail(let id):
-          if let productType = sampleTypes.first(where: { $0.id == id }) {
-              ProductTypeDetailView(productType: productType)
-          }
-      case .feedItem(let id):
-          VStack {
-              Text("Feed Item \(id)")
-                  .font(.title)
-          }
-      case .favoriteItem(let id):
-          VStack {
-              Text("Favorite Item \(id)")
-                  .font(.title)
-          }
-      default:
-          VStack {
-              Text("Tab: \(destination.tab.title)")
-                  .font(.headline)
-              Text("Destination: \(destination.id)")
-                  .font(.subheadline)
-          }
-      }
-      */
-    }
   }
 }
 
@@ -339,268 +242,11 @@ struct AppSplitDetailView: View {
   }
 #endif
 
-// MARK: - Improved Settings View
-
-struct ImprovedSettingsView: View {
-  let captureService: CaptureService?
-  let designSystem: DesignSystemSetup?
-
-  @State private var demoModeEnabled = false
-  @State private var autoSyncEnabled = false
-  @State private var contributeToAI = true
-  @State private var showingUpgradeSheet = false
-  @State private var showingDatabaseResetAlert = false
-
-  init(captureService: CaptureService? = nil, designSystem: DesignSystemSetup? = nil) {
-    self.captureService = captureService
-    self.designSystem = designSystem
-  }
-
-  var body: some View {
-    List {
-      Section {
-        HStack {
-          Image(systemName: "person.circle.fill")
-            .font(.title2)
-            .foregroundColor(.accentColor)
-
-          VStack(alignment: .leading) {
-            Text("Demo User")
-              .font(.headline)
-            Text("demo@thepia.com")
-              .font(.caption)
-              .foregroundColor(.secondary)
-          }
-
-          Spacer()
-
-          Button("Edit") {
-            // Edit account
-          }
-          .buttonStyle(.bordered)
-          .controlSize(.small)
-        }
-      } header: {
-        Text("Account")
-      }
-
-      Section {
-        HStack {
-          VStack(alignment: .leading) {
-            HStack {
-              Text("Free Plan")
-                .font(.headline)
-              Spacer()
-            }
-            Text("Basic recording and local storage")
-              .font(.caption)
-              .foregroundColor(.secondary)
-          }
-
-          Button("Upgrade") {
-            showingUpgradeSheet = true
-          }
-          .buttonStyle(.borderedProminent)
-          .controlSize(.small)
-        }
-      } header: {
-        Text("Plan & Billing")
-      }
-
-      Section {
-        HStack {
-          Label("iCloud Sync", systemImage: "icloud.and.arrow.up")
-          Spacer()
-          Toggle("", isOn: $autoSyncEnabled)
-            .disabled(true)  // Disabled for free tier
-        }
-
-        Button("Sync Now") {
-          // Manual sync
-        }
-        .disabled(true)
-
-        NavigationLink("iCloud Debug") {
-          SimpleiCloudDebugView()
-        }
-
-      } header: {
-        Text("Sync & Backup")
-      } footer: {
-        Text("iCloud sync is available with Premium plan. Debug view shows sync status and logs.")
-      }
-
-      Section {
-        HStack {
-          Label("Contribute to AI Training", systemImage: "brain")
-          Spacer()
-          Toggle("", isOn: $contributeToAI)
-            .disabled(true)  // Free tier always contributes
-        }
-
-        Button("Privacy Policy") {
-          // Show privacy policy
-        }
-
-      } header: {
-        Text("Privacy & Data")
-      } footer: {
-        Text("Free tier recordings help improve our AI. Upgrade to Premium for privacy controls.")
-      }
-
-      Section {
-        HStack {
-          Label("Demo Mode", systemImage: "play.circle")
-          Spacer()
-          Toggle("", isOn: $demoModeEnabled)
-        }
-
-        if demoModeEnabled {
-          Button("Reset Demo Data") {
-            // Reset demo data
-          }
-        }
-
-      } header: {
-        Text("Demo Mode")
-      } footer: {
-        if demoModeEnabled {
-          Text("Demo mode limits database modifications and disables cloud sync.")
-        } else {
-          Text("Enable demo mode to explore the app with sample data.")
-        }
-      }
-
-      #if DEBUG
-        Section {
-          // Database Controls
-          NavigationLink("Database Debug") {
-            DatabaseDebugView()
-          }
-
-          Button("Backup Database") {
-            AppDatasource.shared.triggerManualBackup()
-          }
-
-          Button("Reset Database") {
-            showingDatabaseResetAlert = true
-          }
-          .foregroundColor(.red)
-
-          // Camera Controls
-          if let captureService = captureService, let designSystem = designSystem {
-            Group {
-              HStack {
-                Text("Camera Stream")
-                Spacer()
-                CameraSwitcher(captureService: captureService, designSystem: designSystem)
-              }
-
-              HStack {
-                Text("Power Mode")
-                Spacer()
-                CameraSubduedSwitcher(captureService: captureService, designSystem: designSystem)
-              }
-
-              VStack(alignment: .leading, spacing: 4) {
-                Text("Capture Service Info")
-                  .font(.headline)
-                CaptureServiceInfo(captureService: captureService)
-              }
-            }
-          }
-
-        } header: {
-          Text("Development")
-        } footer: {
-          Text("Development tools for debugging and testing. Camera controls moved from sidebar.")
-        }
-      #endif
-
-      Section {
-        HStack {
-          Text("Version")
-          Spacer()
-          Text("1.0.0")
-            .foregroundColor(.secondary)
-        }
-
-        Button("Help & Support") {
-          // Show help
-        }
-      } header: {
-        Text("About")
-      }
-    }
-    .frame(maxWidth: .infinity, maxHeight: .infinity)
-    .navigationTitle("Settings")
-    #if os(iOS)
-      .navigationBarTitleDisplayMode(.large)
-    #endif
-    .sheet(isPresented: $showingUpgradeSheet) {
-      SimpleUpgradeView()
-    }
-    .alert("Reset Database", isPresented: $showingDatabaseResetAlert) {
-      Button("Cancel", role: .cancel) {}
-      Button("Reset", role: .destructive) {
-        // Reset database
-      }
-    } message: {
-      Text("This will permanently delete all your data and reset the app to its initial state.")
-    }
-  }
-}
-
-struct SimpleUpgradeView: View {
-  @Environment(\.dismiss) private var dismiss
-
-  var body: some View {
-    NavigationView {
-      VStack(spacing: 24) {
-        Image(systemName: "crown.fill")
-          .font(.system(size: 60))
-          .foregroundColor(.yellow)
-
-        Text("Upgrade to Premium")
-          .font(.largeTitle)
-          .fontWeight(.bold)
-
-        Text("Unlock advanced features and cloud sync")
-          .font(.title3)
-          .foregroundColor(.secondary)
-          .multilineTextAlignment(.center)
-
-        Spacer()
-
-        Button("Start Free Trial") {
-          dismiss()
-        }
-        .buttonStyle(.borderedProminent)
-        .controlSize(.large)
-
-        Text("$4.99/month after 7-day free trial")
-          .font(.caption)
-          .foregroundColor(.secondary)
-      }
-      .padding()
-      #if os(iOS)
-        .navigationBarTitleDisplayMode(.inline)
-      #endif
-      .toolbar {
-        ToolbarItem(placement: .confirmationAction) {
-          Button("Close") {
-            dismiss()
-          }
-        }
-      }
-    }
-  }
-}
-
 #if DEBUG
   struct AppSplitView_Previews: PreviewProvider {
     static var previews: some View {
       @StateObject var assetsViewModel = AssetsViewModel()
+
       AppSplitView(
         columnVisibility: .constant(.detailOnly),
         lifecycleView: .constant(.record),
@@ -613,19 +259,21 @@ struct SimpleUpgradeView: View {
         Text("Detail Content")
       }
       .environment(\.assetsViewModel, assetsViewModel)
-      .previewDisplayName("Detail Only")
-    }
-  }
+      .previewDisplayName("Collapsed")
 
-  struct ImprovedSettingsView_Previews: PreviewProvider {
-    static var previews: some View {
-      NavigationView {
-        ImprovedSettingsView(
-          captureService: CaptureService(),
-          designSystem: .light
-        )
+      AppSplitView(
+        columnVisibility: .constant(.automatic),
+        lifecycleView: .constant(.record),
+        path: .constant(NavigationPath()),
+        evidenceViewModel: .createDefault(),
+        cameraViewModel: CameraViewModel(),
+        captureService: CaptureService(),
+        designSystem: .light
+      ) {
+        Text("Detail Content")
       }
-      .previewDisplayName("Settings")
+      .environment(\.assetsViewModel, assetsViewModel)
+      .previewDisplayName("Expanded")
     }
   }
 #endif
